+++ src/pupil_labs/realtime_api/__init__.py
@@ -8,8 +8,11 @@ from .streaming import (
     RTSPData,
     RTSPGazeStreamer,
     RTSPRawStreamer,
+    RTSPVideoFrameStreamer,
+    VideoFrame,
     receive_gaze_data,
     receive_raw_rtsp_data,
+    receive_video_frames,
 )
 from .version import __version__, __version_info__
 
@@ -21,7 +24,10 @@ __all__ = [
     "GazeData",
     "receive_gaze_data",
     "receive_raw_rtsp_data",
+    "receive_video_frames",
     "RTSPData",
     "RTSPGazeStreamer",
     "RTSPRawStreamer",
+    "RTSPVideoFrameStreamer",
+    "VideoFrame",
 ]