+++ src/pupil_labs/realtime_api/streaming/__init__.py
@@ -1,11 +1,15 @@
 from .base import RTSPData, RTSPRawStreamer, receive_raw_rtsp_data
 from .gaze import GazeData, RTSPGazeStreamer, receive_gaze_data
+from .video import RTSPVideoFrameStreamer, VideoFrame, receive_video_frames
 
 __all__ = [
     "GazeData",
     "receive_gaze_data",
     "receive_raw_rtsp_data",
+    "receive_video_frames",
     "RTSPData",
     "RTSPGazeStreamer",
     "RTSPRawStreamer",
+    "RTSPVideoFrameStreamer",
+    "VideoFrame",
 ]